記事一覧とページングのセットをコンポーネント化
